requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap5==2025.6
cryptography==45.0.4
daphne==4.2.1
dj-database-url==3.0.0
Django==5.1.3
django-allauth==65.9.0
django-axes==8.0.0
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-js-asset==3.1.2
djangorestframework==3.16.0
drf-yasg==1.21.10
exceptiongroup==1.3.0
gunicorn==23.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
kombu==5.5.4
MetaTrader5==5.0.4993
numpy==2.2.5
packaging==24.2
pillow==11.0.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pytest==8.4.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.2
Telethon==1.40.0
tomli==2.2.1
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
zope.interface==7.2
